Status App Reorg and Layout Improvments for mobile. Hack and Slash.
