+++ nucypher/network/status_app/assets/favicon.ico
