+++ nucypher/network/status_app/assets/nucypher_logo.png
